fix bug

+++ index.js
@@ -143,7 +143,10 @@ client.on('guildMemberAdd', (member) => {
 
 client.on('message', (message) => {
     if(!message.channel.nsfw) {
-        if(badwords.some(word => message.content.toLocaleLowerCase().includes(word))){
+        if(message.content === '>hentai') {
+            return;
+        }
+        if(badwords.some(word => message.content.toLocaleLowerCase().includes(word) )){
             message.delete()
             message.say(':x: Vous ne pouvez utiliser ces mots seulement dans un salon NSFW !').then(async(resultmessage) => {
                 setTimeout(() => {
